Fix MSVC warnings in OGR core
git-svn-id: https://svn.osgeo.org/gdal/trunk@31597 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogr_srsnode.cpp
@@ -421,7 +421,7 @@ OGRErr OGR_SRSNode::exportToWkt( char ** ppszResult ) const
 
 {
     char        **papszChildrenWkt = NULL;
-    int         nLength = strlen(pszValue)+4;
+    size_t      nLength = strlen(pszValue)+4;
     int         i;
 
 /* -------------------------------------------------------------------- */
@@ -483,7 +483,7 @@ OGRErr OGR_SRSNode::exportToPrettyWkt( char ** ppszResult, int nDepth ) const
 
 {
     char        **papszChildrenWkt = NULL;
-    int         nLength = strlen(pszValue)+4;
+    size_t      nLength = strlen(pszValue)+4;
     int         i;
 
 /* -------------------------------------------------------------------- */
